challenge-2/requirements.txt
open-clip-torch==2.20.0
pandas==2.0.3
pillow==10.0.1
scikit-learn==1.3.0
torch==2.0.1
numpy==1.24.3
matplotlib==3.7.2
tqdm==4.65.0
pathlib==1.0.1
jupyter==1.0.0
